More seperation of concerns/organizing
