+++ public/assets/css/general.css
@@ -1,0 +1,233 @@
+/* CSS VARIABLES */
+
+:root {
+  /* Text Colors */
+  --dark-text: #444444;
+  --dark-text-muted: rgba(68, 68, 68, 0.6);
+  --dark-text-very-muted: rgba(68, 68, 68, 0.1);
+  --dark-text-hidden: rgba(68, 68, 68, 0);
+  --light-text: #f8f8f8;
+  --light-text-muted: rgba(248, 248, 248, 0.6);
+  --light-text-very-muted: rgba(248, 248, 248, 0.1);
+  --light-text-hidden: rgba(248, 248, 248, 0);
+
+  /* Background Colors */
+  --dark-background: #444444;
+  --dark-background-alt: rgba(68, 68, 68, 0.95);
+  --light-background: #f8f8f8;
+  --light-background-alt: rgba(248, 248, 248, 0.95);
+}
+
+/* GENERAL */
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  letter-spacing: 0.4px;
+}
+
+html {
+  /* Percentage of user's browser font-size setting */
+  font-size: 62.5%;
+  scroll-behavior: smooth;
+}
+
+body {
+  font-size: 1.6rem;
+  font-family: 'Merriweather', serif;
+  background-color: var(--light-background);
+  color: var(--dark-text);
+}
+
+p,
+li {
+  line-height: 1.6;
+}
+
+.container {
+  opacity: 0;
+  transition: opacity 1s;
+}
+
+.background-transition {
+  transition: background-color 0.3s;
+}
+
+.fade-in {
+  opacity: 1;
+}
+
+.muted {
+  opacity: 0.7;
+}
+
+.hidden {
+  display: none !important;
+  margin: 0 !important;
+  padding: 0 !important;
+}
+
+.capitalized {
+  text-transform: capitalize;
+}
+
+.btn {
+  border-radius: 4px;
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 1.2rem 2rem;
+  cursor: pointer;
+
+  color: var(--light-text);
+  background-color: var(--dark-background);
+  text-decoration: none;
+
+  box-shadow: rgba(0, 0, 0, 0.1) 0px 3px 25px 0px;
+  transition: box-shadow 1s;
+}
+
+.btn--full {
+  width: 100%;
+}
+
+.btn:disabled {
+  opacity: 0.4;
+}
+
+.btn:hover:not([disabled]) {
+  background-color: var(--dark-background-alt);
+  box-shadow: rgba(0, 0, 0, 0.2) 0px 5px 25px 0px;
+  transition: box-shadow 1s;
+}
+
+.btn--outline {
+  background-color: rgba(0, 0, 0, 0);
+  border: 1px solid var(--dark-text);
+  color: var(--dark-text);
+  opacity: 0.8;
+}
+
+.btn--outline:hover:not([disabled]) {
+  background-color: rgba(0, 0, 0, 0);
+  opacity: 1;
+}
+
+.btn i {
+  margin-right: 0.8rem;
+}
+
+.btn--text {
+  display: inline-block;
+  position: relative;
+  border-radius: 0;
+  color: var(--dark-text);
+  background: none;
+  box-shadow: none;
+  padding: 0;
+  margin: 0;
+  bottom: -1px;
+  border-bottom: 1px solid var(--dark-text);
+  transition: border-bottom 0.6s;
+}
+
+.btn--text:hover:not([disabled]) {
+  background: none;
+  box-shadow: none;
+  border-bottom: 1px solid var(--dark-text-hidden);
+  transition: border-bottom 0.6s;
+}
+
+.btn--icon {
+  display: inline-block;
+  color: var(--dark-text);
+  background: none !important;
+  box-shadow: none !important;
+  padding: 0;
+  margin: 0;
+  opacity: 0.4;
+  transition: opacity 0.3s;
+}
+
+.btn--icon i {
+  margin: 0;
+}
+
+.btn--icon span {
+  margin-left: 0.4rem;
+}
+
+.btn--icon:hover {
+  background: none !important;
+  box-shadow: none !important;
+  opacity: 0.8;
+  transition: opacity 0.3s;
+}
+
+.container {
+  width: 96rem;
+  margin: 0 auto;
+  padding: 1.2rem 3.2rem;
+}
+
+/* HEADER */
+
+header.container {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  color: var(--dark-text-muted);
+  padding: 2rem 0;
+  margin-bottom: 2rem;
+  border-bottom: 1px solid var(--dark-text-very-muted);
+}
+
+.header-text {
+  display: flex;
+  align-items: center;
+  gap: 1.2rem;
+}
+
+.header-text i {
+  opacity: 0.2;
+}
+
+#btnToggleDarkMode {
+  position: absolute;
+  align-self: center;
+  right: 2rem;
+}
+
+/* FOOTER */
+footer.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  color: var(--dark-text-muted);
+  gap: 1.2rem;
+  font-size: 1.2rem;
+  padding: 2rem 0;
+  margin-top: 4rem;
+  display: flex;
+  border-top: 1px solid var(--dark-text-very-muted);
+}
+
+footer.container > ul {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+footer.container > ul > li:not(:last-child):after {
+  content: '|';
+  margin: 0 1.2rem;
+  opacity: 0.4;
+}
+
+footer .icon-footer {
+  width: 2rem;
+  opacity: 0.8;
+}
