+++ public/assets/css/pages/realm.css
@@ -1,0 +1,187 @@
+/* CONTROLS */
+
+section.container--controls {
+  padding: 8rem 2.4rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+#controls h1 {
+  text-align: center;
+  font-size: 6.4rem;
+}
+
+#controls p {
+  text-align: center;
+  font-size: 1.8rem;
+  margin-bottom: 4rem;
+}
+
+#btnStart {
+  margin-bottom: 0.8rem;
+}
+
+.control-buttons {
+  display: flex;
+  gap: 0.8rem;
+  margin-bottom: 1.2rem;
+  width: 48rem;
+}
+
+.hover-label {
+  position: relative;
+  top: 0.4rem;
+  opacity: 0;
+  color: var(--dark-text-muted);
+  font-size: 1.2rem !important;
+  transition: all 0.3s;
+}
+
+/* FAVORITES */
+
+.container.container--favorites p {
+  text-align: center;
+  margin-bottom: 1.2rem;
+}
+
+#favorites {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+  gap: 0.8rem;
+  padding: 0 3.2rem;
+  width: 100%;
+}
+
+.favorite-btn-box {
+  display: flex;
+  justify-content: center;
+  padding: 1.2rem;
+}
+
+#btnFavorite > i.fa-solid {
+  color: #ff8989;
+}
+
+.favorite-badge {
+  display: flex;
+  align-items: center;
+  gap: 0.8rem;
+  border-radius: 55px;
+  padding: 0.2rem 1.2rem;
+  border: 1px solid var(--dark-text-very-muted);
+}
+
+.favorite-badge i {
+  margin: 0;
+}
+
+/* HERO */
+
+#hero {
+  height: 32rem;
+
+  position: relative;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
+
+  color: white;
+}
+
+.ornament {
+  width: 4rem;
+  opacity: 0.6;
+}
+
+.overlay-top-left {
+  position: absolute;
+  top: 0.2rem;
+  left: 0.2rem;
+}
+
+.overlay-top-right {
+  position: absolute;
+  top: 0.2rem;
+  right: 0.2rem;
+}
+
+.overlay-bottom-left {
+  position: absolute;
+  bottom: 0.2rem;
+  left: 0.2rem;
+}
+
+.overlay-bottom-right {
+  position: absolute;
+  bottom: 0.2rem;
+  right: 0.2rem;
+}
+
+.rotate-90 {
+  transform: rotate(90deg);
+}
+.rotate-180 {
+  transform: rotate(180deg);
+}
+.rotate-270 {
+  transform: rotate(270deg);
+}
+
+#hero h2 {
+  margin-bottom: 1.2rem;
+  font-size: 6.4rem;
+  font-family: 'Lobster Two', cursive;
+  font-weight: 400;
+}
+
+/* CONTENT */
+.container--content {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 2.4rem;
+}
+
+article {
+  display: flex;
+  flex-direction: column;
+  padding: 3.2rem;
+  border: 1px dashed var(--dark-text-muted);
+  box-shadow: rgba(0, 0, 0, 0.1) 0px 2px 5px 0px;
+}
+
+article.big {
+  grid-row-end: span 2;
+}
+
+article h2 {
+  margin-bottom: 2.4rem;
+}
+
+article li {
+  margin-bottom: 1.8rem;
+}
+
+article li h3 {
+  opacity: 0.6;
+  font-size: 1.4rem;
+  text-transform: uppercase;
+  letter-spacing: 1.2px;
+  margin-bottom: 0.4rem;
+}
+
+article li i {
+  color: var(--dark-text);
+  margin-right: 0.4rem;
+}
+
+.under-construction {
+  opacity: 0.2;
+}
