+++ public/realm.html
@@ -28,7 +28,9 @@
     />
 
     <!-- Custom CSS -->
-    <link rel="stylesheet" href="assets/css/styles.css" />
+    <link rel="stylesheet" href="assets/css/general.css" />
+    <link rel="stylesheet" href="assets/css/pages/realm.css" />
+    <link rel="stylesheet" href="assets/css/heraldry.css" />
     <link rel="stylesheet" href="assets/css/queries.css" />
 
     <!-- Favicon -->
@@ -98,9 +100,9 @@
     </section>
 
     <!-- HERO -->
-    <section class="container">
+    <section class="container container-hero">
       <div id="hero">
-        <h2><span class="name"></span></h2>
+        <h2></h2>
         <img
           class="overlay-top-left ornament"
           src="assets/images/decorative/button-ornament.svg"
