+++ src/util/Rand.ts
