+++ src/util/Util.ts
@@ -1,9 +1,5 @@
-import Realm from './realm/Realm';
-import Lang from './toponymy/Language';
-import * as numberWords from './number-words.json';
-import * as loremWords from './lorem-words.json';
-import Rand from './Rand';
-
+import Realm from '../realm/Realm';
+import Lang from './Lang';
 export default class Util {
   public static isDarkMode: boolean = false;
 
@@ -118,7 +114,7 @@ export default class Util {
 
   // Tweet a realm
   static shareByTweet(realm: Realm) {
-    let tweet: string = `Explore ${Util.capitalize(
+    let tweet: string = `Explore ${Lang.capitalize(
       Lang.readWord(realm.realmName.name)
     )}, a ${realm.size} ${realm.parentEntity.adjective} ${
       realm.government.rank
@@ -131,30 +127,5 @@ export default class Util {
         tweet,
       '_blank'
     );
-  }
-
-  // Capitalize first letter
-  static capitalize(str: string): string {
-    return str.charAt(0).toUpperCase() + str.substring(1);
-  }
-
-  // Returns any number lower than 20 as a word ('one', 'two', ... 'nineteen')
-  static wordFromNumber(n: number): string {
-    return n < numberWords.length ? numberWords[n] : n.toString();
-  }
-
-  // Quick and dirty placeholder text
-  static lorem() {
-    let str: string = '';
-    for (let i: number = 1; i < 3; i++) {
-      const words: number = Rand.between(3, 10);
-      let arr: string[] = [];
-      for (let j: number = 0; j < words; j++) {
-        arr.push(Rand.pick(loremWords));
-      }
-      str += [Util.capitalize(Rand.pick(loremWords)), ...arr].join(' ') + '. ';
-    }
-
-    return str;
   }
 }
