+++ src/util/seed-words.json
